+++ docs/HISTORY.rst
@@ -6,6 +6,8 @@ Changelog
 ----------------
 
 **CLOSED ISSUES**
+
+- https://github.com/collective/plone.jsonapi.routes/issues/34: Failed POST request return HTTP 200
 - https://github.com/collective/plone.jsonapi.routes/issues/35: DataManager does not check field permissions